Change model for test to Protheus
Iris doesn't have TensorRT models published, but prob does. Allows
for testing that TensorRT works if the GPU device is capable.

Tip: to compare RT vs non-RT variants of the model, just temporarily
change the capabilities in the prob-3.json file to some other values.
This will result in the VAI engine not finding a match for the current
GPU, causing it to falling back to a fp16/32 model variant instead.