Add a zero-fuss benchmark
Allows users to just clone, build and run the benchmark without having
to much around with model caching and authentication.

Purpose is to have a dead-simple reproducible test on all machine
configurations.

Note: Absence of auth.tpz file means the video will have the watermark
overlay added, but that's irrelevant for this purpose. GPU and CPU
efficiency measurement is all that matters in this case.